fix: Update current progress at new track load
There was a race condition between selecting a new track and the track
progress updating; this caused the app to sometimes panic when
selecting a track which has a total duration less than the current
progress as it tries to report a negative remaining time in the playbar
(but it tries this with unsigned values). This commit simply reset
current track progress to 0 when a new track is selected.