Replace black and white default colors with reset. (#742)
Previously, text would be white, which is not visible on bright color
schemes. The color for non-highlighted text now uses the default
foreground color of the users terminal emulator, making it visible in
any theme. The only visual change in dark color schemes is that the
playbar text might go from black to white, depending on the users
terminal emulator (some terminal emulators swap fore- and background
colors on a colored background depending on the background color; this
is IMO the correct behavior anyway).